[0.5.0-23] Change building and activity descriptions, fix critical bugs from 0.5.0-22, improve H4XScreen, adjust some geometry
C++:
-fix a bug in hero.cpp that was enabling decreasing stress below 0

QML/JS:
-add second background to H4XScreen
-fix critical bugs(typos) in HeroesLists and HeroesListScripts
-fix a typo in HeroesListDelegates and HeroView (in setCA fncs)
-change current activity descriptions in HeroView (*still need some C++ fncs for it to work completely)
-adjust geometry and font's pixel size in HeroView and BuildingListDelegate

Data:
-completely change building descriptions
-fix a typo in MadelainLacroix.xml

+++ hero.cpp
@@ -635,7 +635,7 @@ void Hero::decreaseStress(unsigned amount) noexcept
     {
         if (m_currentAttributesValues.stress-amount<m_currentAttributesValues.stressBorder && isStressBorderEffectActive())
         {
-            if (m_currentAttributesValues.stress-amount>0)
+            if (m_currentAttributesValues.stress-static_cast<int>(amount)>0)
                 m_currentAttributesValues.stress-=amount;
             else
                 m_currentAttributesValues.stress=0;
@@ -643,7 +643,7 @@ void Hero::decreaseStress(unsigned amount) noexcept
         }
         else
         {
-            if (m_currentAttributesValues.stress-amount>0)
+            if (m_currentAttributesValues.stress-static_cast<int>(amount)>0)
                 m_currentAttributesValues.stress-=amount;
             else
                 m_currentAttributesValues.stress=0;
